add bottom navigation

+++ app/src/main/java/com/guidoroos/recepten/db/ItemType.kt
@@ -8,5 +8,5 @@ data class ItemType  (
     @PrimaryKey(autoGenerate = true)
     val id : Long = 0,
     val name : String,
-    val resourceId: Int
+    val resourceId: Long
 )
